console: Fix event split open button placement
