+++ pkg/webui/console/containers/event-split-frame/event-split-frame.styl
@@ -14,7 +14,6 @@
 
 .container
   background-color: var(--c-bg-neutral-bold)
-  width: 100%
   position: absolute
   bottom: 0
   left: 0
@@ -50,8 +49,9 @@
   flex-grow: 1
 
 .open-button
-  position: absolute
+  position: fixed
   right: 0
-  top: -3.5rem
+  bottom: 0
   padding: $cs.xxs
+  width: fit-content
