+++ pkg/webui/console/views/app/app.styl
@@ -35,11 +35,3 @@
   width: 100%
   position: absolute
   bottom: 0
-
-.open-button
-  position: sticky
-  right: 0
-  bottom: 0
-  padding: $cs.xxs
-  display: flex
-  justify-content: flex-end
