+++ pkg/webui/locales/en.json
@@ -521,6 +521,7 @@
   "console.containers.documentation-dashboard-panel.index.tts": "The Things Stack",
   "console.containers.documentation-dashboard-panel.index.reference": "Reference",
   "console.containers.documentation-dashboard-panel.index.gettingStarted": "Getting started",
+  "console.containers.event-split-frame.index.expandEventPanel": "Expand live data overlay",
   "console.containers.freq-plans-select.utils.warning": "Frequency plans unavailable",
   "console.containers.freq-plans-select.utils.none": "Do not set a frequency plan",
   "console.containers.freq-plans-select.utils.selectFrequencyPlan": "Select a frequency plan...",
@@ -920,7 +921,6 @@
   "console.views.admin-packet-broker.messages.gatewayVisibilityInformation": "You can use the checkboxes to control what information of your gateways will be visible. Note that this information will be visible to the public and not only to registered networks.",
   "console.views.admin-packet-broker.messages.defaultGatewayVisibilitySet": "Default gateway visibility set",
   "console.views.admin-packet-broker.messages.packetBrokerStatusPage": "Packet Broker Status page",
-  "console.views.app.index.expandEventPanel": "Expand live data overlay",
   "console.views.application-add.index.appDescription": "Within applications, you can register and manage end devices and their network data. After setting up your device fleet, use one of our many integration options to pass relevant data to your external services.{break}Learn more in our guide on <Link>Adding Applications</Link>.",
   "console.views.application-integrations-lora-cloud.index.loraCloudInfoText": "Lora Cloud provides value added APIs that enable simple solutions for common tasks related to LoRaWAN networks and LoRa-based devices. You can setup our LoRaCloud integrations below.",
   "console.views.application-integrations-lora-cloud.index.officialLoRaCloudDocumentation": "Official LoRa Cloud documentation",
