+++ pkg/webui/locales/ja.json
@@ -521,6 +521,7 @@
   "console.containers.documentation-dashboard-panel.index.tts": "",
   "console.containers.documentation-dashboard-panel.index.reference": "",
   "console.containers.documentation-dashboard-panel.index.gettingStarted": "",
+  "console.containers.event-split-frame.index.expandEventPanel": "",
   "console.containers.freq-plans-select.utils.warning": "",
   "console.containers.freq-plans-select.utils.none": "",
   "console.containers.freq-plans-select.utils.selectFrequencyPlan": "",
@@ -920,7 +921,6 @@
   "console.views.admin-packet-broker.messages.gatewayVisibilityInformation": "チェックボックスを使用して、ゲートウェイの情報を表示するように制御することができます。この情報は、登録されたネットワークだけでなく、一般にも公開されることに注意してください",
   "console.views.admin-packet-broker.messages.defaultGatewayVisibilitySet": "デフォルトゲートウェイの可視化設定",
   "console.views.admin-packet-broker.messages.packetBrokerStatusPage": "Packet Broker ステータスページ",
-  "console.views.app.index.expandEventPanel": "",
   "console.views.application-add.index.appDescription": "アプリケーション内では、エンドデバイスとそのネットワークデータを登録、管理することができます。デバイス群を設定した後、多くの統合オプションの1つを使用して、関連データを外部サービスに渡します。<Link>Adding Applications</Link>　に関するガイドで詳しく説明しています",
   "console.views.application-integrations-lora-cloud.index.loraCloudInfoText": "LoRaWANネットワークやLoRaベースのデバイスに関連する一般的なタスクのシンプルなソリューションを可能にする付加価値の高いAPIをローラクラウドが提供します。以下からLoRaCloudのインテグレーションを設定することができます",
   "console.views.application-integrations-lora-cloud.index.officialLoRaCloudDocumentation": "オフィシャル LoRa Cloud ドキュメンテーション",
